refactor(step-9): extract 9 note-domain tools to tools/note-tools.ts, delete utils.ts
The architectural crown of the refactor. Creates src/tools/note-tools.ts
exporting registerNoteTools(server), which registers all 9 note-domain
tools verbatim from main.ts:
- bear-open-note, bear-create-note, bear-search-notes
- bear-add-text, bear-replace-text, bear-add-file
- bear-find-untagged-notes, bear-add-tag, bear-archive-note

All schemas, descriptions, annotations, handler bodies, and registration
order preserved. handleNoteTextUpdate moves into this file as a
module-private helper (used only by bear-add-text and bear-replace-text).

main.ts shrinks from ~970 lines to ~220 — it now creates the McpServer,
calls registerNoteTools(server), then still holds the 3 tag tools
inline (move in step 10). Unused imports in main.ts removed.

src/utils.ts is now empty of content and deleted. Both legacy cycles
(utils <-> notes, utils <-> bear-urls) are gone. Verified:

  npx madge --circular --extensions ts src/  →  ✔ No circular dependency found!

37 tests still pass.
